Add WebSharpJs module.
This allows a CLR module loaded from Electron or Node to script Node.js functions that is initially called from Electron or Node.js.  This is different from the double_edge implementation in that the double_edge allows scripting of Node.js from a stand-alone CLR program that is not loaded from Electron or Node.js scripting.