Example of context menu using Menu and MenuItem
A context, or right-click, menu can be created with the <code>Menu</code> and <code>MenuItem</code> modules as well. You can right-click anywhere in this app to see an example context menu.

The example uses `Menu` and multiple instances of `MenuItem`s created through the use of the strongly typed `MenuItemOptions` class passed as a paramenter to the `MenuItems`'s `Create` instance method.

The menu itself demonstrates the use of sub menus using predefined `Roles` from the `View` menu item as well as assigning a call back to the `Click` menu items.